requirements.txt
psutil==6.1.1
configparser
flask
json
shutil
